refactor: relocate logic to ClassLexer
