+++ lib/entity.js
@@ -1,4 +1,4 @@
-import Lexer from './lexer';
+import { Lexer, ClassLexer } from './lexer';
 
 export default class Entity {
   constructor(el) {
@@ -166,7 +166,8 @@ export default class Entity {
 
   _sanitizeContentExpression(expr, options = {}) {
     if (options.isClass || expr.includes("el.")) {
-      const lexer = new Lexer(expr, options);
+      const LexerClass = options.isClass ? ClassLexer : Lexer;
+      const lexer = new LexerClass(expr, options);
   
       if (expr.includes("el.")) {
         lexer.replace({
@@ -210,29 +211,10 @@ export default class Entity {
     const expr = this.element.getAttribute(':class');
     if (!expr) return;
 
-    const output = this._sanitizeContentExpression(expr, { isClass: true });
-    let newClassNames = [...this.baseClasses];
-
-    if (typeof output === 'string')
-      newClassNames.concat(output.split(' '));
-    
-    if (Array.isArray(output)) {
-      output.forEach((conditional) => {
-        const condition = eval(conditional.test);
-        const consequent = eval(conditional.consequent).split(' ');
-        const alternate = eval(conditional.alternate).split(' ');
-  
-        if (condition) {
-          newClassNames = newClassNames.concat(consequent);
-          newClassNames = newClassNames.filter((value) => !alternate.includes(value));
-        } else {
-          newClassNames = newClassNames.concat(alternate);
-          newClassNames = newClassNames.filter((value) => !consequent.includes(value));
-        }
-      });
-    }
-
-    this.element.className = newClassNames.join(' ');
+    this.element.className = this._sanitizeContentExpression(expr, {
+      base: this.baseClasses,
+      isClass: true,
+    });
   }
 
   evaluateLoadEvents() {
