+++ lib/lexer.js
@@ -32,7 +32,7 @@ function setMemberIdentifier(node, members = []) {
   }
 }
 
-export default class Lexer {
+export class Lexer {
   static debug = false;
   static ID_TYPE = {
     declared: 'declared',
@@ -284,4 +284,37 @@ export default class Lexer {
     
     return output;
   }
+}
+
+export class ClassLexer extends Lexer {
+  constructor(code, options) {
+    super(code, { ...options, isClass: true });
+    this._baseClasses = options.base ?? [];
+  }
+
+  output() {
+    const output  = super.output();
+    let newClassNames = [...this._baseClasses];
+
+    if (typeof output === 'string')
+      newClassNames.concat(output.split(' '));
+
+    if (Array.isArray(output)) {
+      output.forEach((conditional) => {
+        const condition = eval(conditional.test);
+        const consequent = eval(conditional.consequent).split(' ');
+        const alternate = eval(conditional.alternate).split(' ');
+  
+        if (condition) {
+          newClassNames = newClassNames.concat(consequent);
+          newClassNames = newClassNames.filter((value) => !alternate.includes(value));
+        } else {
+          newClassNames = newClassNames.concat(alternate);
+          newClassNames = newClassNames.filter((value) => !consequent.includes(value));
+        }
+      });
+    }
+
+    return [...new Set(newClassNames)].join(' ');
+  } 
 }
