+++ lib/main.js
@@ -1,6 +1,6 @@
 import Entity from './entity'
 import MiniArray from './array';
-import Lexer from './lexer';
+import { Lexer } from './lexer';
 
 let nativeProps = Object.getOwnPropertyNames(window);
